Modified app.R
Modified text in all of the 3 main tools to add explanations, checkboxes to alert user to what to pay attention to with the data (Note: these don't update anything in the tools). I also fixed the labeling of the parameters throughout the VRE and added more detail on what these are.